rework portgen-go to use 'go list -m all' for grabbing hard deps
Modify portgen to call `go list -m all` in addition to `go mod graph`.

This gets the full dependnecy tree and lets us prune the list from graph down to
just the .mod files.

This commit has an issue with the lists being non-sorted and also (in the case
of graph) contains duplicates.

I am not sure how to dedupe yet.